Ex. 4.21 done

+++ part4/BlogList/controllers/blogs.js
@@ -3,13 +3,7 @@ const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
-/*const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}*/
+
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -32,14 +26,13 @@ blogsRouter.get('/:id', (request, response, next) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  //const token = getTokenFrom(request)
+  
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!request.token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
   const user = await User.findById(decodedToken.id)
-  //const users = await User.find({})
-  //const user = await User.findById(users[0]._id)
+ 
 
   const blog = new Blog({
     title: body.title,
@@ -61,8 +54,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'Missing or invalid token' })
+  }
+  const userid = decodedToken.id
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === userid.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
       response.status(204).end()
+  } else {
+      response.status(400).json({ error: 'token or user does not exist' })
+  }
+
     
 })
